Refactor TokenStream extension methods and update package structure
Moved `TokenStream` extension methods (`asFlow` and `asReplyFlow`) to a new `TokenStreamExtensions` file under the `service` package for better modularity. Updated related imports and adjusted the package structure to reflect these changes.
