+++ langchain4j-kotlin/src/main/kotlin/me/kpavlov/langchain4j/kotlin/model/adapters/TokenStreamToStringFlowAdapter.kt
@@ -1,9 +1,9 @@
-package me.kpavlov.langchain4j.kotlin.adapters
+package me.kpavlov.langchain4j.kotlin.model.adapters
 
 import dev.langchain4j.service.TokenStream
 import dev.langchain4j.spi.services.TokenStreamAdapter
 import kotlinx.coroutines.flow.Flow
-import me.kpavlov.langchain4j.kotlin.model.chat.asFlow
+import me.kpavlov.langchain4j.kotlin.service.asFlow
 import java.lang.reflect.ParameterizedType
 import java.lang.reflect.Type
